# backend cost pro/05_openapi.yaml
openapi: 3.1.0
info:
  title: API de Presupuestos de Obra
  version: 1.0.0
  description: |
    API derivada de las historias de usuario del DOC-02 y de las decisiones de
    diseno adoptadas en el DOC-03.

    Patrones transversales:
      - Confirmacion de impacto en dos fases (dryRun -> X-Confirmation-Token)
      - Reversibilidad con undoToken y endpoint generico POST /undo/{token}
      - Concurrencia optimista con ETag e If-Match
      - Validacion en dos capas: 422 para errores duros, arreglo `avisos` para blandos
      - Analitica calculada bajo demanda, con computedAt y ETag
  contact:
    name: Equipo de producto
servers:
  - url: https://api.example.com/v1
    description: Produccion
  - url: https://api.staging.example.com/v1
    description: Pruebas
...
security:
  - bearerAuth: []
paths:
...
  /catalog/apus:
    get:
      tags: [Catalogo APU]
      summary: Listar APUs del catalogo
      description: HU-09. Busqueda insensible a mayusculas y tildes.
      parameters:
        - $ref: '#/components/parameters/q'
        - name: capituloId
          in: query
          schema: { type: string }
        - name: origen
          in: query
          schema:
            type: string
            enum: [BASE, PERSONALIZADO, GENERADO_IA, IMPORTADO]
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Listado paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Apu' }
                  pagina: { $ref: '#/components/schemas/Pagina' }
                  total: { type: integer }
    post:
      tags: [Catalogo APU]
      summary: Crear un APU
      description: HU-10. Requiere perfil admin_catalogo.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApuInput' }
      responses:
        '201':
          description: APU creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Apu' }
        '403': { $ref: '#/components/responses/SinPermiso' }
  /catalog/apus/{apuId}:
    parameters:
      - $ref: '#/components/parameters/apuId'
    get:
      tags: [Catalogo APU]
      summary: Ficha del APU con sus componentes
      description: HU-09
      responses:
        '200':
          description: APU
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Apu' }
    patch:
      tags: [Catalogo APU]
      summary: Editar el APU del catalogo global
      description: |
        HU-11. Alcance de empresa. Requiere admin_catalogo y confirmacion de impacto.
      parameters:
        - $ref: '#/components/parameters/dryRun'
        - $ref: '#/components/parameters/ConfirmationToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApuInput' }
      responses:
        '200':
          description: APU actualizado o analisis de impacto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Impacto' }
        '403': { $ref: '#/components/responses/SinPermiso' }
        '428': { $ref: '#/components/responses/ConfirmacionRequerida' }
  /catalog/apus/{apuId}/duplicate:
    parameters:
      - $ref: '#/components/parameters/apuId'
    post:
      tags: [Catalogo APU]
      summary: Duplicar un APU
      description: HU-12. El nombre de la copia es obligatorio y debe diferir del original.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [descripcion]
              properties:
                descripcion: { type: string }
      responses:
        '201':
          description: Copia creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Apu' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /catalog/apus/{apuId}/impact:
    parameters:
      - $ref: '#/components/parameters/apuId'
    get:
      tags: [Catalogo APU]
      summary: Proyectos y presupuestos afectados por un cambio
      description: HU-11
      responses:
        '200':
          description: Impacto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Impacto' }
  /catalog/counters:
    get:
      tags: [Catalogo APU]
      summary: Fuente unica de los contadores de catalogo
      description: |
        HU-09, decision H-13. Ninguna cifra de catalogo se calcula en el cliente
        ni se escribe a mano en la interfaz.
      parameters:
        - name: projectId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Contadores
          content:
            application/json:
              schema:
                type: object
                properties:
                  apusCatalogo: { type: integer, examples: [3250] }
                  apusEnProyecto: { type: integer, examples: [30] }
                  apusPersonalizados: { type: integer, examples: [12] }
                  insumosCatalogo: { type: integer, examples: [120] }
                  computedAt: { type: string, format: date-time }
...
  /catalog/apus/import:
    post:
      tags: [Catalogo APU]
      summary: Previsualizar una importacion de APUs
      description: |
        HU-13. No aplica nada: devuelve el resumen y un jobId que se confirma aparte.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                archivo: { type: string, format: binary }
      responses:
        '200':
          description: Previsualizacion
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string }
                  nuevos: { type: integer }
                  actualizados: { type: integer }
                  conError: { type: integer }
  /catalog/imports/{jobId}/confirm:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Catalogo APU]
      summary: Aplicar el lote previsualizado
      description: |
        HU-13. Atomico: se aplica todo el lote valido o no se aplica nada.
        Genera automaticamente un punto de restauracion previo.
      responses:
        '200':
          description: Importacion aplicada
          content:
            application/json:
              schema:
                type: object
                properties:
                  aplicados: { type: integer }
                  restorePointId: { type: string }
  /projects/{projectId}/analytics/summary:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Analitica]
      summary: Panel principal del proyecto
      description: |
        HU-17, decision H-14. Se calcula al solicitarlo. No existe endpoint de
        refresco obligatorio. Devuelve computedAt y ETag.
      responses:
        '200':
          description: Resumen calculado
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResumenPanel' }
        '304':
          description: Sin cambios desde el ETag enviado
  /projects/{projectId}/analytics/cost-intelligence:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Analitica]
      summary: Costo por metro cuadrado y composicion del gasto
      description: HU-18. Si el proyecto no tiene area, los indicadores por m2 vienen en null.
      responses:
        '200':
          description: Indicadores
          content:
            application/json:
              schema:
                type: object
                properties:
                  computedAt: { type: string, format: date-time }
                  costoDirectoPorM2: { $ref: '#/components/schemas/Dinero' }
                  recursos:
                    type: array
                    items:
                      type: object
                      properties:
                        tipo: { type: string, enum: [MATERIALES, MANO_OBRA, EQUIPOS, TRANSPORTE] }
                        valor: { $ref: '#/components/schemas/Dinero' }
                        pct: { type: number }
                        porM2: { $ref: '#/components/schemas/Dinero' }
  /projects/{projectId}/analytics/consolidated:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Analitica]
      summary: Consolidado de recursos del proyecto
      description: |
        HU-19. Recorre todas las actividades y agrega por insumo. El total de cada
        grupo debe cuadrar exactamente con el rubro del panel principal.
      parameters:
        - name: grupo
          in: query
          required: true
          schema:
            type: string
            enum: [MATERIAL, MANO_OBRA, EQUIPO, TRANSPORTE]
      responses:
        '200':
          description: Consolidado
          content:
            application/json:
              schema:
                type: object
                properties:
                  computedAt: { type: string, format: date-time }
                  total: { $ref: '#/components/schemas/Dinero' }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        insumoId: { type: string }
                        descripcion: { type: string }
                        unidad: { type: string }
                        cantidadTotal: { $ref: '#/components/schemas/Cantidad' }
                        valorUnitario: { $ref: '#/components/schemas/Dinero' }
                        valorTotal: { $ref: '#/components/schemas/Dinero' }
  /projects/{projectId}/analytics/alerts:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Analitica]
      summary: Desviaciones y avisos accionables
      description: HU-18, decision H-16. Alimenta la franja de avisos del panel principal.
      responses:
        '200':
          description: Alertas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        codigo:
                          type: string
                          enum: [PRECIO_DESACTUALIZADO, CANTIDAD_ATIPICA, APU_INCOMPLETO, COTIZACION_PENDIENTE]
                        mensaje: { type: string }
                        severidad: { type: string, enum: [INFO, ADVERTENCIA, ERROR] }
                        recurso:
                          type: object
                          properties:
                            tipo: { type: string }
                            id: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
      examples:
        default: { value: proj_9f2a }
...
    apuId:
      name: apuId
      in: path
      required: true
      schema: { type: string }
...
    q:
      name: q
      in: query
      description: Busqueda de texto, insensible a mayusculas y tildes
      schema: { type: string }
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    cursor:
      name: cursor
      in: query
      description: Cursor opaco devuelto por la pagina anterior
      schema: { type: string }
    dryRun:
      name: dryRun
      in: query
      description: Si es true no aplica nada y devuelve solo el analisis de impacto
      schema: { type: boolean, default: false }
    IfMatch:
      name: If-Match
      in: header
      description: ETag del recurso, para control de concurrencia optimista
      schema: { type: string }
...
    ConfirmationToken:
      name: X-Confirmation-Token
      in: header
      description: Token devuelto por el analisis de impacto previo
      schema: { type: string }
  responses:
...
    SinPermiso:
      description: El perfil del usuario no alcanza para esta operacion
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflicto:
      description: El estado actual impide la operacion
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ConfirmacionRequerida:
      description: Falta el X-Confirmation-Token de una operacion con impacto
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Dinero:
      type: string
      description: Cadena decimal con dos decimales. Nunca coma flotante.
      pattern: '^-?\d+\.\d{2}$'
      examples: ['58696983.45']
    Cantidad:
      type: string
      description: Cadena decimal con hasta cuatro decimales
      examples: ['26.0400']
    Pagina:
      type: object
      properties:
        limit: { type: integer }
        siguienteCursor: { type: string, nullable: true }
        hayMas: { type: boolean }
...
    Error:
      type: object
      properties:
        tipo:
          type: string
          enum: [PETICION_INVALIDA, NO_AUTENTICADO, SIN_PERMISO, NO_ENCONTRADO,
                 CONFLICTO, EXPIRADO, CONFLICTO_VERSION, VALIDACION,
                 CONFIRMACION_REQUERIDA, CUOTA_AGOTADA, DEMASIADAS_PETICIONES,
                 SERVICIO_NO_DISPONIBLE]
        mensaje: { type: string, description: Apto para mostrar al usuario }
        traceId: { type: string }
        errores:
          type: array
          items:
            type: object
            properties:
              campo: { type: string }
              codigo: { type: string }
              esperado: { type: string }
...
    Componente:
      type: object
      properties:
        insumoId: { type: string }
        descripcion: { type: string }
        grupo:
          type: string
          enum: [MATERIAL, MANO_OBRA, EQUIPO, TRANSPORTE]
        unidad: { type: string }
        rendimiento: { $ref: '#/components/schemas/Cantidad' }
        costoUnitario: { $ref: '#/components/schemas/Dinero' }
        subtotal: { $ref: '#/components/schemas/Dinero' }
    ApuInput:
      type: object
      required: [descripcion, unidad, capituloId]
      properties:
        descripcion: { type: string, maxLength: 250 }
        unidad: { type: string }
        capituloId: { type: string }
        componentes:
          type: array
          items:
            type: object
            required: [insumoId, rendimiento]
            properties:
              insumoId: { type: string }
              rendimiento: { type: number, exclusiveMinimum: 0 }
    Apu:
      type: object
      properties:
        id: { type: string }
        codigo: { type: string }
        descripcion: { type: string }
        unidad: { type: string }
        capitulo:
          type: object
          properties:
            id: { type: string }
            nombre: { type: string }
        costoUnitario:
          $ref: '#/components/schemas/Dinero'
        desglose:
          type: object
          properties:
            materiales: { $ref: '#/components/schemas/Dinero' }
            manoObra: { $ref: '#/components/schemas/Dinero' }
            equipo: { $ref: '#/components/schemas/Dinero' }
        componentes:
          type: array
          items: { $ref: '#/components/schemas/Componente' }
        origen:
          type: string
          enum: [BASE, PERSONALIZADO, GENERADO_IA, IMPORTADO]
        version: { type: integer }
...
    Impacto:
      type: object
      properties:
        proyectosActivos: { type: integer }
        presupuestosBorrador: { type: integer }
        presupuestosAprobados: { type: integer }
        variacionUnitaria: { $ref: '#/components/schemas/Dinero' }
        confirmationToken: { type: string }
...
    ResumenPanel:
      type: object
      properties:
        computedAt: { type: string, format: date-time }
        estadoVacio: { type: boolean }
        actividades: { type: integer }
        costoDirecto: { $ref: '#/components/schemas/Dinero' }
        aiu:
          type: object
          properties:
            valor: { $ref: '#/components/schemas/Dinero' }
            pct: { type: number }
        total: { $ref: '#/components/schemas/Dinero' }
        costoPorM2:
          oneOf:
            - $ref: '#/components/schemas/Dinero'
            - type: 'null'
        recursos:
          type: array
          items:
            type: object
            properties:
              tipo: { type: string }
              valor: { $ref: '#/components/schemas/Dinero' }
              pct: { type: number }
              porM2: { $ref: '#/components/schemas/Dinero' }
        capitulos:
          type: array
          description: Ordenados por participacion descendente por el servidor
          items:
            type: object
            properties:
              id: { type: string }
              nombre: { type: string }
              valor: { $ref: '#/components/schemas/Dinero' }
              pct: { type: number }
              actividades: { type: integer }
...